doas-pratham: force checkout the dotfiles repo
When doing 'git checkout' for the dotfiles repo, the '.bashrc' file is
already present (curtsy of the 'bash' package on Arch Linux and I reckon
on other distros too). Therefore, use the '-f' flag with the 'git checkout'
command.

Since this is a fresh installation, the force checkout shouldn't hurt.